Fix issue where unit tests do not reference relative file paths from the same path as executables like terminal_client.go
